+++ sequencer.go
@@ -6,7 +6,10 @@ import (
 
 type Sequencer struct {
 	Record `capid:"skip"`
-	Index  uint64 `capid:"0"`
+
+	// TODO: serialisation of uint64 is broken.
+	// Update cap'n'proto, change to uint64 in future.
+	Index int64 `capid:"0"`
 }
 
 func (seq *Sequencer) BucketName() string {
@@ -23,5 +26,5 @@ func (seq *Sequencer) SetKey(sId string) error {
 
 func (seq *Sequencer) Next() uint64 {
 	seq.Index++
-	return seq.Index
+	return uint64(seq.Index)
 }
